+++ zikzak_inappwebview_internal_annotations/src/exchangeable_object_property.dart
@@ -1,0 +1,6 @@
+class ExchangeableObjectProperty {
+  final Function? serializer;
+  final Function? deserializer;
+
+  const ExchangeableObjectProperty({this.serializer, this.deserializer});
+}