+++ zikzak_inappwebview/CHANGELOG.md
@@ -1,3 +1,13 @@
+## 1.7.0 - 2025-04-26
+
+* added internal anootations for next release
+* Updated dependencies to use hosted references
+
+## 1.5.0 - 2025-04-26
+
+* support compileSdk 35 fix android issues
+* Updated dependencies to use hosted references
+
 ## 1.1.0
 
 - Updated all subpackages to version 1.1.0